fix: Move CharacterBuilderContext to separate file to resolve CI linting errors
- Extract context definition to src/contexts/CharacterBuilderContext.ts
- Update imports in provider and hook files
- Remove unused imports to pass linting
- Fixes react-refresh/only-export-components linting rule
