+++ src/components/index.ts
@@ -7,6 +7,7 @@ export * from './TraitBadge';
 
 // Character Builder Context
 export * from '../character/creation/CharacterBuilderContext';
+export * from '../contexts/CharacterBuilderContext';
 export * from '../hooks/useCharacterBuilder';
 
 // Existing components
